test with RPLCD from pypi
