+++ MANIFEST.in
@@ -1,2 +1,2 @@
-include cbpi4-LCDisplay/config.yaml
-recursive-include cbpi4-LCDisplay/static *
+include cbpi4-LCD/config.yaml
+recursive-include cbpi4-LCD/static *
