+++ cbpi4-LCD/config.yaml
@@ -1,0 +1,2 @@
+name: cbpi4-LCD
+version: 4
