--- cbpi4-LCDisplay/RPLCD/__init__.py
@@ -1,14 +1,0 @@
-import warnings
-
-from .contextmanagers import cursor, cleared  # noqa
-
-
-__version__ = '1.3.0'
-
-
-class CharLCD:
-    def __new__(cls, *args, **kwargs):
-        from .gpio import CharLCD as GpioCharLCD
-        warnings.warn("Using RPLCD.CharLCD directly is deprecated. " +
-                      "Use RPLCD.gpio.CharLCD instead!", DeprecationWarning)
-        return GpioCharLCD(*args, **kwargs)
