--- cbpi4-LCDisplay/RPLCD/common.py
@@ -1,119 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Copyright (C) 2013-2018 Danilo Bargen
-
-Permission is hereby granted, free of charge, to any person obtaining a copy of
-this software and associated documentation files (the "Software"), to deal in
-the Software without restriction, including without limitation the rights to
-use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
-the Software, and to permit persons to whom the Software is furnished to do so,
-subject to the following conditions:
-
-The above copyright notice and this permission notice shall be included in all
-copies or substantial portions of the Software.
-
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
-FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
-COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
-IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
-CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
-
-"""
-from __future__ import print_function, division, absolute_import, unicode_literals
-
-import itertools
-import time
-
-
-# # # BIT PATTERNS # # #
-
-# Commands
-LCD_CLEARDISPLAY = 0x01
-LCD_RETURNHOME = 0x02
-LCD_ENTRYMODESET = 0x04
-LCD_DISPLAYCONTROL = 0x08
-LCD_CURSORSHIFT = 0x10
-LCD_FUNCTIONSET = 0x20
-LCD_SETCGRAMADDR = 0x40
-LCD_SETDDRAMADDR = 0x80
-
-# Flags for display entry mode
-LCD_ENTRYRIGHT = 0x00
-LCD_ENTRYLEFT = 0x02
-LCD_ENTRYSHIFTINCREMENT = 0x01
-LCD_ENTRYSHIFTDECREMENT = 0x00
-
-# Flags for display on/off control
-LCD_DISPLAYON = 0x04
-LCD_DISPLAYOFF = 0x00
-LCD_CURSORON = 0x02
-LCD_CURSOROFF = 0x00
-LCD_BLINKON = 0x01
-LCD_BLINKOFF = 0x00
-
-# Flags for display/cursor shift
-LCD_DISPLAYMOVE = 0x08
-LCD_CURSORMOVE = 0x00
-
-# Flags for display/cursor shift
-LCD_DISPLAYMOVE = 0x08
-LCD_CURSORMOVE = 0x00
-LCD_MOVERIGHT = 0x04
-LCD_MOVELEFT = 0x00
-
-# Flags for function set
-LCD_8BITMODE = 0x10
-LCD_4BITMODE = 0x00
-LCD_2LINE = 0x08
-LCD_1LINE = 0x00
-LCD_5x10DOTS = 0x04
-LCD_5x8DOTS = 0x00
-
-# Flags for RS pin modes
-RS_INSTRUCTION = 0x00
-RS_DATA = 0x01
-
-
-# # # Helper classes # # #
-
-class Alignment(object):
-    left = LCD_ENTRYLEFT
-    right = LCD_ENTRYRIGHT
-
-
-class ShiftMode(object):
-    cursor = LCD_ENTRYSHIFTDECREMENT
-    display = LCD_ENTRYSHIFTINCREMENT
-
-
-class CursorMode(object):
-    hide = LCD_CURSOROFF | LCD_BLINKOFF
-    line = LCD_CURSORON | LCD_BLINKOFF
-    blink = LCD_CURSOROFF | LCD_BLINKON
-
-
-# # # HELPER FUNCTIONS # # #
-
-def msleep(milliseconds):
-    """Sleep the specified amount of milliseconds."""
-    time.sleep(milliseconds / 1000.0)
-
-
-def usleep(microseconds):
-    """Sleep the specified amount of microseconds."""
-    time.sleep(microseconds / 1000000.0)
-
-
-def sliding_window(seq, lookahead):
-    """
-    Create a sliding window with the specified number of lookahead characters.
-    """
-    it = itertools.chain(iter(seq), ' ' * lookahead)  # Padded iterator
-    window_size = lookahead + 1
-    result = tuple(itertools.islice(it, window_size))
-    if len(result) == window_size:
-        yield result
-    for elem in it:
-        result = result[1:] + (elem,)
-        yield result
