--- cbpi4-LCDisplay/config.yaml
@@ -1,2 +1,0 @@
-name: cbpi4-LCDisplay
-version: 4
