+++ setup.py
@@ -7,9 +7,9 @@ this_directory = path.abspath(path.dirname(__file__))
 with open(path.join(this_directory, 'README.md'), encoding='utf-8') as f:
     long_description = f.read()
 
-setup(name='cbpi4-LCDisplay',
-      version='0.0.13.a7',
-      description='CraftBeerPi4 LCD Plugin Mod',
+setup(name='cbpi4-LCD',
+      version='0.1.0.a1',
+      description='CraftBeerPi4 LCD Plugin',
       author='Alexander Vollkopf',
       author_email='avollkopf@web.de',
       url='https://github.com/PiBrewing/cbpi4-LCDisplay',
@@ -21,10 +21,10 @@ setup(name='cbpi4-LCDisplay',
         # If any package contains *.txt or *.rst files, include them:
       '': ['*.txt', '*.rst', '*.yaml'],
       'LCDisplay': ['*','*.txt', '*.rst', '*.yaml']},
-      #packages=['cbpi4-LCDisplay'],
-      #install_requires=[
-      #'cbpi4>=4.1.10.rc2'
-      #],
+      packages=['cbpi4-LCD'],
+      install_requires=[
+      'RPLCD'
+      ],
       long_description=long_description,
       long_description_content_type='text/markdown'
      )
